Add files via upload
Step 1:
In step one we can open the Microchip Studio to write a code in  C language for controlling LED light with a push button over an output port of ATmega32.
Step 2:
In step two we can build the given code by pressing F7. In the comment section, we can see the program build successfully.
Step 3:
In step three we can run the program. 
After that we pause the program and we can see the Program Memory.

Simulate:
Step 1:
In step four now we can open Proteus and draw the circuit.
We can select ATmega32 IC, a resister, a push button, ground, and LED light. 
And make the connection.
Step 2:
By clicking ATmega32 IC a new dialogue box appears.
Now we can choose the Hexa file of PROGRAM as we write in Microchip Studio. 
And press the OK button after selecting a file.
In the end, now we run it.
By Pressing the push button, the LED will be ON.